Fix build failures on Win64 since r14215.
R=svenpanne@chromium.org

Review URL: https://codereview.chromium.org/14075006

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14221 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/global-handles.cc
@@ -588,7 +588,7 @@ bool GlobalHandles::IterateObjectGroups(ObjectVisitor* v,
   int surviving_info_index = 0;
 
   UniqueId current_group_id(0);
-  size_t current_group_start = 0;
+  int current_group_start = 0;
   bool any_group_was_visited = false;
 
   for (int i = 0; i <= object_groups_.length(); ++i) {
